+++ src/ProtocolZoo/show.jl
@@ -55,3 +55,34 @@ function Base.show(io::IO, m::MIME"text/html", p::EntanglementConsumer)
     </div>
     """)
 end
+
+function Base.show(io::IO, m::MIME"text/html", p::BellPairSampler)
+    sampledpairs = length(p._log)
+    total_time = sampledpairs > 0 ? last(p._log).t : 0.0
+    average_time_between_pairs = sampledpairs > 0 ? total_time / sampledpairs : 0.0
+    average_fidelity = sampledpairs > 0 ? sum(x -> x.fidelity, p._log) / sampledpairs : 0.0
+    average_zz = sampledpairs > 0 ? sum(x -> x.zz, p._log) / sampledpairs : 0.0
+    average_xx = sampledpairs > 0 ? sum(x -> x.xx, p._log) / sampledpairs : 0.0
+    average_yy = sampledpairs > 0 ? sum(x -> x.yy, p._log) / sampledpairs : 0.0
+    print(io,
+    """
+    <div class="quantumsavory_show quantumsavory_protocol quantumsavory_protocol_bell_pair_sampler">
+    <h1><code class="quantumsavory_typename quantumsavory_protocol_typename">BellPairSampler</code> protocol</h1>
+    <address>on nodes $(compactstr(p.net[p.nodeA])) and $(compactstr(p.net[p.nodeB]))</address>
+    <dl>
+    <dt>Sampled pairs</dt>
+    <dd>$(sampledpairs)</dd>
+    <dt>Total time</dt>
+    <dd>$(total_time)</dd>
+    <dt>Average time between pairs</dt>
+    <dd>$(average_time_between_pairs)</dd>
+    <dt>Average Bell fidelity estimate</dt>
+    <dd>$(average_fidelity)</dd>
+    <dt>Average stabilizers ZZ | XX | YY</dt>
+    <dd>$(average_zz) | $(average_xx) | $(average_yy)</dd>
+    </dl>
+    <h2>Log</h2>
+    $(pretty_table(String, p._log, column_labels=["Time", "ZZ", "XX", "YY", "Fidelity"], formatters=[PrettyTables.fmt__printf("%5.3f")], backend=:html, maximum_number_of_rows=25))
+    </div>
+    """)
+end